requirements.txt
# pip install -r requirements.txt
flask
rembg
pillow